use .cc extension to the test file
working on db queries
