+++ Makefile
@@ -27,5 +27,5 @@ infer: infer.cc $(OBJ)
 mpi_lda: mpi_lda.cc $(OBJ)
 	$(MPICC) $(CFLAGS) $(OBJ) $< -o $@
 
-db_test: db_test.cpp $(OBJ)
+db_test: db_test.cc $(OBJ)
 	$(CC) $(CFLAGS) -lpqxx $(OBJ) $< -o $@
