+++ db_test.cc
@@ -1,0 +1,50 @@
+#include <iostream>
+#include <string>
+#include <pqxx/pqxx>
+
+#define STATUS_LDA_ANALYSIS 5
+
+using namespace std;
+using namespace pqxx;
+
+int main(int argc, char* argv[]) {
+    std::string const base_sql = "SELECT * from topic_modeling_request where pk=? and status=?";
+    int pk = 4;
+    unsigned long pos;
+    try {
+        connection C("dbname = lm_backend user = lm_admin password = 1qazxsw2 \
+      hostaddr = 127.0.0.1 port = 5432");
+        if (C.is_open()) {
+            cout << "Opened database successfully: " << C.dbname() << endl;
+        } else {
+            cout << "Can't open database" << endl;
+            return 1;
+        }
+        /* Create SQL statement */
+        std::string sql = base_sql;
+        pos = sql.find('?');
+        sql.insert(pos, std::to_string(pk));
+        pos = sql.find('?');
+        sql.insert(pos, std::to_string(STATUS_LDA_ANALYSIS));
+
+        /* Create a non-transactional object. */
+        nontransaction N(C);
+
+        /* Execute SQL query */
+        result R( N.exec( sql ));
+
+        /* List down all the records */
+        for (result::const_iterator c = R.begin(); c != R.end(); ++c) {
+            cout << "ID = " << c[0].as<int>() << endl;
+            cout << "Name = " << c[1].as<string>() << endl;
+            cout << "Age = " << c[2].as<int>() << endl;
+            cout << "Address = " << c[3].as<string>() << endl;
+            cout << "Salary = " << c[4].as<float>() << endl;
+        }
+        cout << "Operation done successfully" << endl;
+        C.disconnect ();
+    } catch (const std::exception &e) {
+        cerr << e.what() << std::endl;
+        return 1;
+    }
+}
